Verbeterde volledige CRUD-functionaliteit voor Animals inclusief validatie, weergave en styling. Ook had ik eerst de .gitignore als text bestand waardoor het problemen veroorzaakte in de commits maar opgelost.

+++ Controllers/AnimalsController.cs
@@ -29,19 +29,13 @@ namespace ZooApp.Controllers
         // GET: Animals/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var animal = await _context.Animals
                 .Include(a => a.Category)
                 .Include(a => a.Enclosure)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (animal == null)
-            {
-                return NotFound();
-            }
+            if (animal == null) return NotFound();
 
             return View(animal);
         }
@@ -49,14 +43,11 @@ namespace ZooApp.Controllers
         // GET: Animals/Create
         public IActionResult Create()
         {
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Id");
-            ViewData["EnclosureId"] = new SelectList(_context.Enclosures, "Id", "Id");
+            PopulateDropdowns();
             return View();
         }
 
         // POST: Animals/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Species,CategoryId,Size,DietaryClass,ActivityPattern,EnclosureId,SpaceRequirement,SecurityRequirement")] Animal animal)
@@ -67,40 +58,28 @@ namespace ZooApp.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Id", animal.CategoryId);
-            ViewData["EnclosureId"] = new SelectList(_context.Enclosures, "Id", "Id", animal.EnclosureId);
+            PopulateDropdowns(animal);
             return View(animal);
         }
 
         // GET: Animals/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var animal = await _context.Animals.FindAsync(id);
-            if (animal == null)
-            {
-                return NotFound();
-            }
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Id", animal.CategoryId);
-            ViewData["EnclosureId"] = new SelectList(_context.Enclosures, "Id", "Id", animal.EnclosureId);
+            if (animal == null) return NotFound();
+
+            PopulateDropdowns(animal);
             return View(animal);
         }
 
         // POST: Animals/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Species,CategoryId,Size,DietaryClass,ActivityPattern,EnclosureId,SpaceRequirement,SecurityRequirement")] Animal animal)
         {
-            if (id != animal.Id)
-            {
-                return NotFound();
-            }
+            if (id != animal.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -111,38 +90,25 @@ namespace ZooApp.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!AnimalExists(animal.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!AnimalExists(animal.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Id", animal.CategoryId);
-            ViewData["EnclosureId"] = new SelectList(_context.Enclosures, "Id", "Id", animal.EnclosureId);
+            PopulateDropdowns(animal);
             return View(animal);
         }
 
         // GET: Animals/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var animal = await _context.Animals
                 .Include(a => a.Category)
                 .Include(a => a.Enclosure)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (animal == null)
-            {
-                return NotFound();
-            }
+            if (animal == null) return NotFound();
 
             return View(animal);
         }
@@ -153,10 +119,7 @@ namespace ZooApp.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var animal = await _context.Animals.FindAsync(id);
-            if (animal != null)
-            {
-                _context.Animals.Remove(animal);
-            }
+            if (animal != null) _context.Animals.Remove(animal);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -165,6 +128,18 @@ namespace ZooApp.Controllers
         private bool AnimalExists(int id)
         {
             return _context.Animals.Any(e => e.Id == id);
+        }
+
+        // Dropdown-helper
+        private void PopulateDropdowns(Animal? animal = null)
+        {
+            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name", animal?.CategoryId);
+            ViewData["EnclosureId"] = new SelectList(_context.Enclosures, "Id", "Name", animal?.EnclosureId);
+
+            ViewBag.Size = new SelectList(Enum.GetValues(typeof(Size)));
+            ViewBag.DietaryClass = new SelectList(Enum.GetValues(typeof(DietaryClass)));
+            ViewBag.ActivityPattern = new SelectList(Enum.GetValues(typeof(ActivityPattern)));
+            ViewBag.SecurityRequirement = new SelectList(Enum.GetValues(typeof(SecurityLevel)));
         }
     }
 }
